refactor: BDD tests - consistent mock component names and move msg svc to separate file
