--- test/bdd/fixtures/didcomm/docker-compose.yml
@@ -1,68 +1,0 @@
-#
-# Copyright SecureKey Technologies Inc. All Rights Reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-version: '2'
-
-services:
-
-  wallet.mock.example.com:
-    container_name: wallet.mock.example.com
-    image: ${ARIES_AGENT_REST_IMAGE}:${ARIES_AGENT_REST_IMAGE_TAG}
-    environment:
-      - ARIESD_API_HOST=${WALLET_AGENT_HOST}:${WALLET_AGENT_API_PORT}
-      - ARIESD_INBOUND_HOST=http@${WALLET_AGENT_HOST}:${WALLET_AGENT_INBOUND_PORT}
-      - ARIESD_INBOUND_HOST_EXTERNAL=http@http://wallet.mock.example.com:${WALLET_AGENT_INBOUND_PORT}
-      - ARIESD_WEBHOOK_URL=http://wallet.mock.example.com
-      - ARIESD_DATABASE_TYPE=mysql
-      - ARIESD_DATABASE_URL=aries:aries-secret-pw@tcp(mysql:3306)/
-      - ARIESD_DATABASE_PREFIX=aries_wallet
-      - ARIESD_DEFAULT_LABEL=aries-wallet
-      - ARIESD_HTTP_RESOLVER=trustbloc@http://did.rest.example.com:8072/1.0/identifiers
-      - ARIESD_WEBHOOK_URL=http://wallet-webhook-mock.example.com:${WALLET_WEBHOOK_PORT}
-    ports:
-      - ${WALLET_AGENT_INBOUND_PORT}:${WALLET_AGENT_INBOUND_PORT}
-      - ${WALLET_AGENT_API_PORT}:${WALLET_AGENT_API_PORT}
-    command: start
-    networks:
-      adapter-rest_bdd_net:
-        aliases:
-          - wallet.mock.example.com
-
-  wallet-webhook-mock.example.com:
-    container_name: wallet-webhook-mock.example.com
-    image: ${MOCK_WEBHOOK_IMAGE}:${MOCK_WEBHOOK_IMAGE_TAG}
-    environment:
-      - WEBHOOK_PORT=${WALLET_WEBHOOK_PORT}
-    ports:
-      - ${WALLET_WEBHOOK_PORT}:${WALLET_WEBHOOK_PORT}
-    networks:
-      adapter-rest_bdd_net:
-        aliases:
-          - wallet-webhook-mock.example.com
-
-  issuer.adapter.mock.example.com:
-    container_name: issuer.adapter.mock.example.com
-    image: ${ARIES_AGENT_REST_IMAGE}:${ARIES_AGENT_REST_IMAGE_TAG}
-    environment:
-      - ARIESD_API_HOST=${ISSUER_AGENT_HOST}:${ISSUER_AGENT_API_PORT}
-      - ARIESD_INBOUND_HOST=http@${ISSUER_AGENT_HOST}:${ISSUER_AGENT_INBOUND_PORT}
-      - ARIESD_INBOUND_HOST_EXTERNAL=http@http://issuer.adapter.mock.example.com:${ISSUER_AGENT_INBOUND_PORT}
-      - ARIESD_DATABASE_TYPE=mysql
-      - ARIESD_DATABASE_URL=aries:aries-secret-pw@tcp(mysql:3306)/
-      - ARIESD_DATABASE_PREFIX=issuer_adapter_mock
-      - ARIESD_DEFAULT_LABEL=issuer-adapter
-      - ARIESD_AUTO_ACCEPT=true
-    ports:
-      - ${ISSUER_AGENT_INBOUND_PORT}:${ISSUER_AGENT_INBOUND_PORT}
-      - ${ISSUER_AGENT_API_PORT}:${ISSUER_AGENT_API_PORT}
-    command: start
-    networks:
-      adapter-rest_bdd_net:
-        aliases:
-          - issuer.adapter.mock.example.com
-
-networks:
-  adapter-rest_bdd_net:
-    external: true
