+++ test/bdd/fixtures/integration/.env
@@ -6,6 +6,25 @@
 
 ISSUER_PORT=9080
 
+# Aries Agent configurations
+ARIES_AGENT_REST_IMAGE=docker.pkg.github.com/trustbloc-cicd/aries-framework-go/agent-rest
+ARIES_AGENT_REST_IMAGE_TAG=amd64-0.1.5-snapshot-dcb7f47
+
+# Webhook configurations
+MOCK_WEBHOOK_IMAGE=docker.pkg.github.com/trustbloc/edge-adapter/mock-webhook
+MOCK_WEBHOOK_IMAGE_TAG=latest
+
+# Wallet Agent configurations
+WALLET_AGENT_HOST=0.0.0.0
+WALLET_AGENT_INBOUND_PORT=9081
+WALLET_AGENT_API_PORT=9082
+WALLET_WEBHOOK_PORT=9083
+
+# Issuer Agent configurations
+ISSUER_AGENT_HOST=0.0.0.0
+ISSUER_AGENT_INBOUND_PORT=10010
+ISSUER_AGENT_API_PORT=10011
+
 # TrustBloc DID Method
 BLOC_DOMAIN=testnet.trustbloc.local
