+++ test/bdd/fixtures/integration/docker-compose.yml
@@ -7,8 +7,63 @@ version: '2'
 
 services:
 
-  issuer.example.com:
-    container_name: issuer.example.com
+  mock-wallet.com:
+    container_name: mock-wallet.com
+    image: ${ARIES_AGENT_REST_IMAGE}:${ARIES_AGENT_REST_IMAGE_TAG}
+    environment:
+      - ARIESD_API_HOST=${WALLET_AGENT_HOST}:${WALLET_AGENT_API_PORT}
+      - ARIESD_INBOUND_HOST=http@${WALLET_AGENT_HOST}:${WALLET_AGENT_INBOUND_PORT}
+      - ARIESD_INBOUND_HOST_EXTERNAL=http@http://mock-wallet.com:${WALLET_AGENT_INBOUND_PORT}
+      - ARIESD_DATABASE_TYPE=mysql
+      - ARIESD_DATABASE_URL=aries:aries-secret-pw@tcp(mysql:3306)/
+      - ARIESD_DATABASE_PREFIX=aries_wallet
+      - ARIESD_DEFAULT_LABEL=aries-wallet
+      - ARIESD_HTTP_RESOLVER=trustbloc@http://did.rest.example.com:8072/1.0/identifiers
+      - ARIESD_WEBHOOK_URL=http://webhook.mock-wallet.com:${WALLET_WEBHOOK_PORT}
+    ports:
+      - ${WALLET_AGENT_INBOUND_PORT}:${WALLET_AGENT_INBOUND_PORT}
+      - ${WALLET_AGENT_API_PORT}:${WALLET_AGENT_API_PORT}
+    command: start
+    networks:
+      adapter-rest_bdd_net:
+        aliases:
+          - mock-wallet.com
+
+  webhook.mock-wallet.com:
+    container_name: webhook.mock-wallet.com
+    image: ${MOCK_WEBHOOK_IMAGE}:${MOCK_WEBHOOK_IMAGE_TAG}
+    environment:
+      - WEBHOOK_PORT=${WALLET_WEBHOOK_PORT}
+    ports:
+      - ${WALLET_WEBHOOK_PORT}:${WALLET_WEBHOOK_PORT}
+    networks:
+      adapter-rest_bdd_net:
+        aliases:
+          - webhook.mock-wallet.com
+
+  mock-issuer-adapter.com:
+    container_name: mock-issuer-adapter.com
+    image: ${ARIES_AGENT_REST_IMAGE}:${ARIES_AGENT_REST_IMAGE_TAG}
+    environment:
+      - ARIESD_API_HOST=${ISSUER_AGENT_HOST}:${ISSUER_AGENT_API_PORT}
+      - ARIESD_INBOUND_HOST=http@${ISSUER_AGENT_HOST}:${ISSUER_AGENT_INBOUND_PORT}
+      - ARIESD_INBOUND_HOST_EXTERNAL=http@http://mock-issuer-adapter.com:${ISSUER_AGENT_INBOUND_PORT}
+      - ARIESD_DATABASE_TYPE=mysql
+      - ARIESD_DATABASE_URL=aries:aries-secret-pw@tcp(mysql:3306)/
+      - ARIESD_DATABASE_PREFIX=issuer_adapter_mock
+      - ARIESD_DEFAULT_LABEL=issuer-adapter
+      - ARIESD_AUTO_ACCEPT=true
+    ports:
+      - ${ISSUER_AGENT_INBOUND_PORT}:${ISSUER_AGENT_INBOUND_PORT}
+      - ${ISSUER_AGENT_API_PORT}:${ISSUER_AGENT_API_PORT}
+    command: start
+    networks:
+      adapter-rest_bdd_net:
+        aliases:
+          - mock-issuer-adapter.com
+
+  mock-issuer.com:
+    container_name: mock-issuer.com
     image: docker.pkg.github.com/trustbloc/edge-adapter/mock-issuer:latest
     environment:
       - ISSUER_PORT=${ISSUER_PORT}
@@ -17,7 +72,7 @@ services:
     networks:
       adapter-rest_bdd_net:
         aliases:
-          - issuer.example.com
+          - mock-issuer.com
 
   governance.vcs.example.com:
     container_name: governance.vcs.example.com
